workflows/workflow-004/pip-requirements.txt
alabaster==0.7.13
blinker==1.9.0
Flask==3.0.2
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
requests==2.32.3
Werkzeug==3.1.4
molscrub==0.2.2
useful-rdkit-utils